requirements.txt
eth_account==0.8.0
nacl==0.0.0
PyNaCl==1.5.0
python-dotenv==1.0.0
requests==2.28.2
